% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisMR_cML.R
\name{cismr_cML_DP}
\alias{cismr_cML_DP}
\title{Main function to run cisMR-cML with data perturbation}
\usage{
cismr_cML_DP(
  b_exp,
  b_out,
  Sig_exp_inv,
  Sig_out_inv,
  K_vec = 0:(length(b_exp) - 2),
  random_start = 0,
  random_start_pert = 0,
  maxit = 200,
  num_pert = 100,
  random_seed = 0,
  min_theta_range = -0.5,
  max_theta_range = 0.5,
  n
)
}
\arguments{
\item{b_exp}{A vector of conditional genetic effect sizes on the exposure.}

\item{b_out}{A vector of conditional genetic effect sizes on the outcome.}

\item{Sig_exp_inv}{The inverse of the covariance matrix of \code{b_exp}.}

\item{Sig_out_inv}{The inverse of the covariance matrix of \code{b_out}.}

\item{K_vec}{Sets of candidate K's, the constraint parameter representing number of invalid IVs.}

\item{random_start}{Number of random start points for cisMR-cML-BIC, default is 0.}

\item{random_start_pert}{Number of random start points for cisMR-cML with data perturbation, default is 0.}

\item{maxit}{Maximum number of iterations for each optimization, default is 200.}

\item{num_pert}{Number of perturbation, default is 100.}

\item{random_seed}{Random seed, an integer. Default is
0, which does not set random seed; user could specify a positive integer
as random seed to get replicable results.}

\item{min_theta_range}{The lower bound for the randomly generated initial causal effect size, default is -0.5.}

\item{max_theta_range}{The upper bound for the randomly generated initial causal effect size, default is 0.5.}

\item{n}{Sample size of exposure or outcome GWAS.}
}
\value{
A list
\describe{
\item{BIC_theta}{Estimated causal effect from cisMR-cML-BIC}
\item{BIC_se}{Estimate standard error for \code{BIC_theta}}
\item{BIC_p}{P-value for \code{BIC_theta}}
\item{BIC_invalid}{Invalid IVs selected by cisMR-cML-BIC}
\item{BIC_DP_theta}{Estimated causal effect from cisMR-cML-DP }
\item{BIC_DP_se}{Estimate standard error for \code{BIC_DP_theta}}
\item{BIC_DP_p}{P-value for \code{BIC_DP_theta}}
}
}
\description{
Main function to run cisMR-cML with data perturbation
}
