% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mr_dat}
\alias{mr_dat}
\title{Effect of LDL cholesterol on coronary artery disease (CAD) on the PCSK9 region}
\format{
A \code{list} with 9 slots.
}
\usage{
mr_dat
}
\description{
Effect of LDL cholesterol on coronary artery disease (CAD) on the PCSK9 region
}
\keyword{datasets}
